feat: add MODLOG for easier logging in modules
MODLOG() is a shortcut for syslog(); prefer it over fprintf().

Note: it does have a draw-back; because it injects into the format,
there must /be/ a format; there is no 2-argument version of MODLOG().

As a supporting addition, expose the module's name to the module
iself. This might be useful for configuration in the future.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>