feat: add basic DPMS status module
If DPMS is available, displays a coffee-cup to signify the screen will
not sleep, or some information about screen saving status. If
possible, it will show the suspend timeout less the idle time—i.e.,
the minutes remaining till the screen will suspend. If not, it will
simply show the three configured timeouts in seconds.

Note: it only uses the suspend timeout for the timer calculation. It
would be much better to use the next non-zero timeout and state which
timeout is upcoming.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>